# Better compaction heuristic