--- a/VERSION
+++ b/VERSION
@@ -1,4 +1,4 @@
-4.13.0+dev0-2020-10-19
+4.13.0+dev1+compact-study
 
 # The version string is the first line of this file.
 # It must be in the format described in stdlib/sys.mli