Compaction is triggered when the current memory overhead becomes too large. This can happen in two cases:
* when fragmentation makes a lot of memory unallocatable
* when the live data has shrunk a lot
In both cases, compacting the heap is a good idea.

Currently, we trigger the compaction based on an indirect computation of the overhead. In practice, this computation is noisy and the estimate is often way off. In this case, the `compact_heap_maybe` function will force a nonincremental major GC (a necessary  first step for compaction) then get the actual overhead (easy to measure because the heap is now clean) and notice that the actual overhead is much lower than the estimate, and will give up on the compaction.

A better estimate of the overhead leads to fewer of these spurious major collections.

In this PR, our better estimate is simply a precise measurement of the overhead at the end of the previous cycle.

How much better is it? That depends on your program, of course. Here are the results from a few entries of the sandmark suite. On the left the old estimate: a plot of the actual overhead (y-axis) against the estimate (x-axis); on the right the new estimate: a plot of the current overhead (y-axis) against the previous cycle's overhead (x-axis). The closer to the diagonal, the better.

Note that this is comparing the two estimates, but the program is still using the old one.

We start with well-behaved programs:

* coq
![coqc-abstract](https://user-images.githubusercontent.com/19104/106744386-fc6db680-661f-11eb-8a76-0afd7f52e7db.jpg)
Note the outliers on the x=5000 line: they are actually at 1 million and were clipped to make the graph readable. In these cases, the estimated value is infinite.
* cubicle 
![cubicle-szymanski](https://user-images.githubusercontent.com/19104/106744734-6ede9680-6620-11eb-9dc5-0d428586726e.jpg)
In this case, compaction is never triggered. Nevertheless, the new estimate looks much better.

We also have some programs with a less-standard use of memory:
* levinson-durbin 
![levinson_durbin](https://user-images.githubusercontent.com/19104/106750311-fc71b480-6627-11eb-922e-ed441727ef71.jpg)
This is numerical computation. The allocation pattern is quite different from coq or cubicle. The old estimate is always wrong: lots of points are estimated at 0 when the actual overhead is more than 1000 and would justify a compaction, while some estimates are at 1000000 (again, clipped at 5000) while the actual overhead is low. The new estimate is not perfect but still much better than the old.
* pidigits5 
![pidigits5](https://user-images.githubusercontent.com/19104/106750481-304cda00-6628-11eb-9e4c-e7dff9ef7ac4.jpg)
This is mostly out-of-heap data (zarith) and nothing in the heap. This program has unusual behavior: it alternates between major GC and compaction, which explains the strange shape of the graph on the right. The new estimate looks better, but doesn't necessarily behave much better.