@gasche 2021-02-03
You have nice visualization for the accuracy of the heuristic, but no data on the impact on performance. Are you planning to report on this later/separately?