@damiendoligez 2021-02-04
Marking this as WIP because @lpw25 pointed out to me that I should be able to avoid the extra instruction in the inner loop of `mark_slice` and @stedolan thinks it could make a noticeable difference because of register pressure. New commit coming soon.