@damiendoligez 2021-03-17
@gasche: here's the performance impact. I've used 4.11.1 with best-fit as the baseline, and a subset of sandmark (eliminating the benchmarks that don't do any compaction).

### Time (lower is better):
![time-norm-1](https://user-images.githubusercontent.com/19104/111481116-f0066e80-8732-11eb-9f72-474c92f6d1f1.png)
I re-ran the benchmark to get an idea of the noise level:
![time-norm-2](https://user-images.githubusercontent.com/19104/111481881-ac603480-8733-11eb-9883-d36fb341ae9e.png)
I'd say everything is noise except the improvement for pidigits and zarith.

### Maximum heap size (Iower is better):
![heap-norm-1](https://user-images.githubusercontent.com/19104/111481515-54293280-8733-11eb-91a8-d2833dc91711.png)
Note that spectralnorm has a tiny heap (only one chunk in the baseline) and what happens is a kind of quantization error: the compaction reduces the size of the heap, which then needs to increase and gets larger than the baseline.

### Two things I find a bit surprising:
* most benchmarks do fewer compactions, but the two that benefit most (pidigits and zarith) get better results by doing a lot more compactions.
* the levinson-durbin benchmark manages to do 2000 major GCs and 700 compactions in less than 3 seconds. Of course its heap is small, but still.

### Raw data:
|   name   |   time (base) | time (new)    | major coll (base) | major coll (new) | compactions (base) | compactions (new) |
|----------|--------------:|--------------:|------------------:|-----------------:|-------------------:|------------------:|
|chameneos_redux_lwt.600000|1.71|1.71|204|201|2|2|
|cpdf.scale|10.86|11.50|27|28|1|0|
|cpdf.squeeze|13.64|13.20|32|32|1|0|
|durand-kerner-aberth.|0.11|0.12|85|83|2|2|
|kb.|3.02|2.98|276|252|27|2|
|kb_no_exc.|1.81|1.79|206|195|13|1|
|levinson-durbin.|2.57|2.58|2125|1906|708|635|
|naive-multilayer.|3.15|3.24|123|141|2|2|
|pidigits5.10_000|5.02|3.14|3137|2996|1045|1354|
|revcomp2.|2.32|2.38|19|17|1|0|
|sequence_cps.10000|1.27|1.28|667|761|2|2|
|setrip.-enc_-rseed_1067894368|1.14|1.14|7|9|2|2|
|spectralnorm2.5_500|6.11|6.28|5|7|0|2|
|test_decompress.64_524_288|3.09|3.12|587|489|96|64|
|yojson_ydump.sample.json|0.59|0.58|19|16|6|5|
|zarith_pi.5000|1.23|0.89|749|750|250|375|