@gasche 2021-03-20
Interesting! Interpreting the data you give is hard because the plots that you had before are measuring both metrics on the old implementation, while the benchmarks you give compare the old to the new implementation, which does its compactions and GCs at different time. Ideally we would see a plot of the memory usage over the program lifetime, with either the old and the new implementations, to be able to understand what is going on.

One remark:

> This is mostly out-of-heap data (zarith) and nothing in the heap.

Is this true? My understanding is that Zarith implements bignums by allocating Custom or Abstract blocks that hold the gmp bigint *inside the OCaml block*, see [caml_z.c : ml_z_alloc](https://github.com/ocaml/Zarith/blob/release-1.9.1/caml_z.c#L322-L333).

The program that is being measured is [zarith_pi.ml](https://github.com/ocaml-bench/sandmark/blob/372b837/benchmarks/zarith/zarith_pi.ml), which computes the decimals of pi by building a rational approximation with larger and larger (relative) numerators and denominators. I think that what is happening is that larger and larger abstract/custom blocks are allocated inside the heap, most of them short-lived, but that the blocks that were freed recently were slightly smaller, creating fragmentation issues indeed.