@gasche 2021-03-20
Here is a made-up story of what might happen in the zarith-pi benchmark:

General context:
- As said earlier: the program allocates short-lived blocks that are larger and larger over time.
- Wild guess: during most of the program life, these blocks go directly to the major heap.
- When an incremental GC slice occurs, many of the blocks freed are smaller than the "current size" of the blocks.
  This increases fragmentation, as these free spaces cannot be used to fit new allocations.
- When a non-incremental GC occurs, almost all the blocks are freed.
- Therefore, if a compaction triggers shortly after a full major GC, then the actual overhead is very small, but if we wait a long time to compact we observe a much larger overhead. On the "actual overhead" plots, we observe a lot of samples with an actual overhead around 2000x (note: the plot values are percentages, right?), this corresponds to the overhead of the maximal delay between a major GC and the following compaction.

Note: which major allocator is used in these measurements? I would expect best-fit to do better than other allocators here, by being more effective at coalescing slightly-too-small blocks into large-enough blocks.

Narrative for the performance difference:
- The previous, broken estimation for this program was 0 most of the time, until it suddenly became infinite. This program had 750 major-GC cycles and 250 compactions, this suggests that we had, regularly, 1 compaction for 3 full GC cycles.
- The new estimate uses the overhead at the end of the previous GC cycle, so it increases gradually over time. With the new implementation, this program has 375 compactions, which suggests that the gradual increase crosses the threshold quicker, leading to a compaction every 2 major cycles.
- Non-incremental major GCs are excellent for this program, which allocates a lot of short-lived blocks on the major heap. Having non-incremental cycles more often results in a large performance improvement.