@gasche 2021-03-20
Remark:

> In this PR, our better estimate is simply a precise measurement of the overhead at the end of the previous cycle.

I'm not familiar with how the incremental major GC works, but is the overhead computed at the end of a GC cycle really "precise"? If I understand correctly, it is precise if we assume that an incremental GC cycle behaves like a non-incremental GC, but it may be imprecise if there was a lot of allocations and deaths in major heap in the middle of the cycle (because in this case the number of marked words may be much lower than the number of dead words). If this is correct, then what you call `actual[0]` is really "ground truth", but what you call `actual[-1]` is not, it is also a sort of estimate.