@lpw25 2021-03-20
I think that counting everything that is marked during a cycle gives you a precise measure of everything that was live at the start of the cycle. It sounds like maybe you are picturing the behaviour of an *incremental update* collector rather than a *snapshot at the beginning* collector.