@gasche 2021-03-20
Ah, so you are saying that the incremental major GC guarantees that all the memory that was live at the start of the cycle will have been traversed at the end of the cycle (and only that memory?), so the overhead measurement at the end of a GC cycle corresponds to the ground truth about the memory not at this point, but at the beginning of the cycle. Fair enough. (This means that there is more than one-major-cycle of distance between the measurement and its use, but we really are talking about two ground truths.)