@damiendoligez 2021-03-23
> My understanding is that Zarith implements bignums by allocating Custom or Abstract blocks

Oh yes that's right. And your explanation of the fragmentation is correct. This is similar to the framgentation you get when you append repeatedly to a large string.